+++ tests/settings.py
@@ -6,7 +6,9 @@ DATABASES = {
     }
 }
 SECRET_KEY = "top_secret"
-USE_TZ = True  # This is here to avoid Django's warning when running tests about the upcoming change in Django 5.0
+# This is here to avoid Django's warning when running tests about the upcoming
+# change in Django 5.0
+USE_TZ = True
 
 DEFAULT_FILE_STORAGE = "cloudflare_images.storage.CloudflareImagesStorage"
 CLOUDFLARE_IMAGES_ACCOUNT_ID = "account_id"